fix up comments
