+++ koku/reporting_common/models.py
@@ -4,41 +4,11 @@
 #
 """Models for shared reporting tables."""
 from django.db import models
-from django.db.models import IntegerChoices
 from django.utils import timezone
 
-from common.enum import StrEnum
-
-
-class ReportStep(IntegerChoices):
-    DOWNLOADING = 3
-    PROCESSING = 4
-    OCP_CLOUD_PROCESSING = (5, "OCP-Cloud-Processing")
-
-
-class Status(IntegerChoices):
-    DONE = 1
-    FAILED = 2
-
-
-class CombinedChoices(IntegerChoices):
-    DOWNLOADING = ReportStep.DOWNLOADING
-    PROCESSING = ReportStep.PROCESSING
-    OCP_CLOUD_PROCESSING = ReportStep.OCP_CLOUD_PROCESSING
-    DONE = Status.DONE
-    FAILED = Status.FAILED
-
-
-class ManifestStep(StrEnum):
-    DOWNLOAD = "download"
-    PROCESSING = "processing"
-    SUMMARY = "summary"
-
-
-class ManifestState(StrEnum):
-    START = "start"
-    END = "end"
-    FAILED = "failed"
+from reporting_common.states import CombinedChoices
+from reporting_common.states import ReportStep
+from reporting_common.states import Status
 
 
 class CostUsageReportManifest(models.Model):
